V1.5.1 - Begun adding equipment for horses
