+++ Part2/Equipment.java
@@ -1,7 +1,17 @@
 class Equipment {
-    int confidenceBoost;
+    private String name;
+    private double confidenceBoost; //multiplier
 
-    public Equipment(int confidenceBoost) {
+    public Equipment(String name, double confidenceBoost) {
+        this.name = name;
         this.confidenceBoost = confidenceBoost;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public double getConfidenceBoost() {
+        return confidenceBoost;
     }
 }
