--- Part2/data/horses.csv
@@ -1,11 +1,0 @@
-H, Horsey,0.5
-H,Horsey,0.5
-J,Jeremy,0.2
-P,PIPPI LONGSTOCKING,0.6
-K,KOKOMO,0.6
-E,EL JEFE,0.4
-a,What,0.2
-M,Marlon,0.9
-a,Test,0.5
-a,Test,0.5
-a,Laughing Horse,0.9
